[src] Create _common for BirdCLEF
